requirements.txt
streamlit 
langchain
langchain_community
langchain-huggingface
google-generativeai
PyPDF2
python-dotenv
sentence-transformers
faiss-cpu


